기능구현 완료

+++ src/main/java/org/example/backendpractice/dao/PostRepository.java
@@ -3,5 +3,5 @@ package org.example.backendpractice.dao;
 import org.example.backendpractice.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, Long> {
 }
